Write rule-set from scratch (~43 score now)
I've had ~60 score with default rules + decorellation + q19
optimization, BUT it became quite hard to improve on the huge set of
default rules (~100). I made a decision that it's gonna be more fun,
control and understanding to assemble the whole ruleset from scratch.
My score went down, but I've already learned a few things in the
process.